refactor: split backend main tests by responsibility
