+++ backend/tests/conftest.py
@@ -1,8 +1,20 @@
 from pathlib import Path
 import sys
 
+import pytest
+from fastapi.testclient import TestClient
 
 PROJECT_ROOT = Path(__file__).resolve().parents[2]
 
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
+
+from backend.main import create_app
+
+
+@pytest.fixture
+def client(tmp_path) -> TestClient:
+    database_url = f"sqlite:///{tmp_path / 'test.db'}"
+
+    with TestClient(create_app(database_url)) as test_client:
+        yield test_client
